Decouples the actual SplatMesh from the Viewer
